Make Readers carry own Record
Instead of relying on the fallback implementation of iterate from BioGenerics,
make the Readers carry their own Record object which is being read into when
iterated.
A new keyword argument `copy` of the Readers then control if the internal Record
or a copy is yielded on iteration.
This simplifies the current while-loop idiom for efficiently iterating over
records in an IO, while also speeding up copied iteration