Make BioSequences a weak dependency
In the upcoming Julia 1.9 release, Julia supports weak depencies which allow
deps to be supported, but not required.
With this change, on Julia 1.9, BioSequences is not automatically loaded when
FASTX is. This allows BioSequences (and its deps) to be not be loaded when FASTX
is, reducing loading time by about 1/3.